Refactor Visualizer class
